Refactor imports for consistency: update BindFile import paths across multiple modules
