+++ CityOfBinds/__init__.py
@@ -1,14 +1,18 @@
 # /CityOfBinds/__init__.py
 from .src.game.binds import Bind, WASDBind, iWASDBind
-from .src.game.bind_file.bind_file import BindFile
+from .src.game import BindFile
 from .src.game.rotating_binds import (
     RotatingBind,
     RandomBinds,
     RandomWalk,
     WASDRotatingBind,
 )
-from .src.content_managers.templates.bind_template import BindTemplate, WASDBindTemplate
-from .src.content_managers.templates.bind_file_template import BindFileTemplate
+from .src.content_managers import (
+    BindTemplate,
+    WASDBindTemplate,
+    BindFileTemplate,
+    BindFileGraph,
+)
 from .src.content_publisher.bfg_publisher import BFGPublisher
 
 __all__ = [
@@ -23,5 +27,6 @@ __all__ = [
     "BindTemplate",
     "WASDBindTemplate",
     "BindFileTemplate",
+    "BindFileGraph",
     "BFGPublisher",
 ]
