+++ CityOfBinds/src/content_managers/graph/bind_file_graph.py
@@ -1,6 +1,6 @@
 import networkx as nx
 import copy
-from ...game.bind_file import BindFile
+from ...game import BindFile
 
 
 trigger_conditions = {
