+++ CityOfBinds/src/content_publisher/bfg_publisher.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 from ..game.binds import Bind
-from ..game.bind_file import BindFile, BindFileConstants
+from ..game import BindFile, BindFileConstants
 from ...utils import StrPath
 from ...utils import _FileGraphPublisher
