+++ CityOfBinds/src/game/__init__.py
@@ -1,0 +1,7 @@
+from .bind_file.bind_file import BindFile
+from .bind_file.constants import BindFileConstants
+
+__all__ = [
+    "BindFile",
+    "BindFileConstants",
+]
