--- CityOfBinds/src/game/bind_file/__init__.py
@@ -1,4 +1,0 @@
-from .bind_file import BindFile
-from .constants import BindFileConstants
-
-__all__ = ["BindFile", "BindFileConstants"]
